Enhance product retrieval and update data models
- Modified `GetAll` method in `ProductController.cs` to use `GetAllIncludingAsync` for fetching related entities.
- Added `ApiSettings` section in `appsettings.Development.json` for API base URL.
- Changed `StockAmount` type from `byte` to `int` in `ProductEntity.cs` and `ProductDto.cs`.
- Restructured migration files, removing large blocks of code in `20250830211321_InitialCreate.Designer.cs` and `20250830211321_InitialCreate.cs`.
- Updated `HomeController.cs` to use `ProductDto` and added error handling for null responses.
- Adjusted `Products.cshtml` view to reflect model changes and improved product display logic.
- Removed `ApiSettings` section from `appsettings.json`.
- Updated `ProductImageDto.cs` to include a `Product` navigation property.
- Modified `ProductListingViewModel.cs` to change `DiscountedPrice` to a standard property.
- Updated `AutoMapperProfiles.cs` for new mappings between `ProductDto` and `ProductListingViewModel`.
- Updated migration files and `init.sql` to reflect new schema and initial data seeding.

+++ AVMAPP.Data/Entities/ProductEntity.cs
@@ -19,7 +19,7 @@ namespace AVMAPP.Data.Entities
         public string Name { get; set; } = "Undefined";
         public decimal Price { get; set; }
         public string Description { get; set; } = "Undefined";
-        public byte StockAmount { get; set; }=0;
+        public int StockAmount { get; set; }=0;
         [JsonIgnore]
         public virtual ICollection<ProductImageEntity> Images { get; set; } = new List<ProductImageEntity>();
         public virtual ICollection<ProductCommentEntity> Comments { get; set; } = new List<ProductCommentEntity>();
